Add retry policy mechanism for HTTP client operations
Introduce a structured retry policy for HTTP client operations across the application. Key changes include:

- Added a `methodName` parameter to retry policy methods for improved logging and tracking.
- Updated `InjectionExtensions.cs` to include a new using directive and configure the HTTP client with a retry policy handler.
- Modified pipeline classes (`CreateBatchNode.cs`, `DownloadFilesNode.cs`, `ProductSearchNode.cs`, `UploadFilesNode.cs`) to instantiate retry policies with specific method names.
- Enhanced `HttpClientPolicyProvider` with overloads for retry policy methods to accept method names.
- Updated `JobService.cs` to include method names in retry policies for better traceability.
- Adjusted unit tests in `HttpClientPolicyProviderTests.cs` to validate the new method name parameter and removed outdated commented-out tests.

These changes improve the robustness and transparency of HTTP client operations.
